Isolate mouse tracker into own module and INDENTATION
