+++ app/javascript/modules/init_tracker.js
@@ -1,0 +1,25 @@
+function initTracker() {
+  let canvas = document.getElementById('grid');
+
+  function getMousePos(canvas, evt) {
+    var rect = canvas.getBoundingClientRect();
+    return {
+      x: evt.clientX - rect.left,
+      y: evt.clientY - rect.top
+    };
+  }
+
+
+  canvas.addEventListener('mousemove', function(evt) {
+    var a = stage.style.transform;
+    var b = a.substring(7);
+    var c = b.slice(0, -1);
+    var d = c.split(',');
+    var mousePos = getMousePos(canvas, evt);
+    var coordsX = Math.floor(mousePos.x / d[0]);
+    var coordsY = Math.floor(mousePos.y / d[0]);
+    console.log("X: " + coordsX + "\n" + "Y: " + coordsY + "\n");
+  }, false);
+}
+
+export { initTracker };
