+++ app/javascript/packs/application.js
@@ -1,6 +1,8 @@
 import "bootstrap";
 import { initGrid } from '../modules/init_grid';
 import { initStage } from '../modules/init_stage';
+import { initTracker } from '../modules/init_tracker';
 initGrid();
 initStage();
+initTracker();
